refactor(model): split apart model used in large simulation module
